Move BedspaceSearch Api Endpoint to CAS3
