+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/controller/cas3/Cas3BedspaceSearchController.kt
@@ -1,8 +1,9 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.controller
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.controller.cas3
+
 import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Service
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.BedsApiDelegate
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.BedSearchResults
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.cas3.BedspacesCas3Delegate
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas3BedspaceSearchResults
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.TemporaryAccommodationBedSearchParameters
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.UserService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas3.Cas3BedspaceSearchService
@@ -10,12 +11,12 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.transformer.cas3.Cas3Bed
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.extractEntityFromCasResult
 
 @Service
-class BedSearchController(
+class Cas3BedspaceSearchController(
   private val userService: UserService,
   private val cas3BedspaceSearchService: Cas3BedspaceSearchService,
   private val cas3BedspaceSearchResultTransformer: Cas3BedspaceSearchResultTransformer,
-) : BedsApiDelegate {
-  override fun bedsSearchPost(temporaryAccommodationBedSearchParameters: TemporaryAccommodationBedSearchParameters): ResponseEntity<BedSearchResults> {
+) : BedspacesCas3Delegate {
+  override fun bedspaceSearchPost(temporaryAccommodationBedSearchParameters: TemporaryAccommodationBedSearchParameters): ResponseEntity<Cas3BedspaceSearchResults> {
     val user = userService.getUserForRequest()
 
     val searchResult = cas3BedspaceSearchService.findBedspaces(
@@ -24,7 +25,7 @@ class BedSearchController(
     )
 
     return ResponseEntity.ok(
-      cas3BedspaceSearchResultTransformer.transformDomainToApi(extractEntityFromCasResult(searchResult)),
+      cas3BedspaceSearchResultTransformer.transformDomainToCas3BedspaceSearchResults(extractEntityFromCasResult(searchResult)),
     )
   }
 }
