+++ src/main/resources/static/api.yml
@@ -2194,6 +2194,7 @@ paths:
           $ref: '_shared.yml#/components/responses/500Response'
   /beds/search:
     post:
+      description: deprecated, use /cas3/bedspaces/search
       summary: Searches for available Beds within the given parameters
       operationId: bedsSearchPost
       requestBody:
