+++ src/main/resources/static/codegen/built-cas3-api-spec.yml
@@ -133,6 +133,35 @@ paths:
           $ref: '#/components/responses/403Response'
         500:
           $ref: '#/components/responses/500Response'
+  /bedspaces/search:
+    post:
+      summary: Searches for available Beds within the given parameters
+      operationId: bedspaceSearchPost
+      requestBody:
+        content:
+          'application/json':
+            schema:
+              $ref: '#/components/schemas/TemporaryAccommodationBedSearchParameters'
+        required: true
+      responses:
+        200:
+          description: successful operation
+          content:
+            'application/json':
+              schema:
+                $ref: '#/components/schemas/Cas3BedspaceSearchResults'
+        400:
+          description: invalid params
+          content:
+            'application/problem+json':
+              schema:
+                $ref: '#/components/schemas/ValidationError'
+        401:
+          $ref: '#/components/responses/401Response'
+        403:
+          $ref: '#/components/responses/403Response'
+        500:
+          $ref: '#/components/responses/500Response'
 components:
   responses:
     401Response:
@@ -4735,3 +4764,74 @@ components:
         - person
         - arrivalDate
         - departureDate
+    Cas3BedspaceSearchResults:
+      type: object
+      properties:
+        resultsRoomCount:
+          type: integer
+          description: How many distinct Rooms the Beds in the results belong to
+        resultsPremisesCount:
+          type: integer
+          description: How many distinct Premises the Beds in the results belong to
+        resultsBedCount:
+          type: integer
+          description: How many Beds are in the results
+        results:
+          type: array
+          items:
+            $ref: '#/components/schemas/Cas3BedspaceSearchResult'
+      required:
+        - resultsRoomCount
+        - resultsPremisesCount
+        - resultsBedCount
+        - results
+    Cas3BedspaceSearchResult:
+      type: object
+      properties:
+        premises:
+          $ref: '#/components/schemas/BedSearchResultPremisesSummary'
+        room:
+          $ref: '#/components/schemas/BedSearchResultRoomSummary'
+        bed:
+          $ref: '#/components/schemas/BedSearchResultBedSummary'
+        overlaps:
+          type: array
+          items:
+            $ref: '#/components/schemas/Cas3BedspaceSearchResultOverlap'
+      required:
+        - premises
+        - room
+        - bed
+        - overlaps
+    Cas3BedspaceSearchResultOverlap:
+      type: object
+      properties:
+        name:
+          type: string
+        crn:
+          type: string
+        sex:
+          type: string
+        personType:
+          $ref: '#/components/schemas/PersonType'
+        days:
+          type: integer
+        bookingId:
+          type: string
+          format: uuid
+        roomId:
+          type: string
+          format: uuid
+        assessmentId:
+          type: string
+          format: uuid
+        isSexualRisk:
+          type: boolean
+      required:
+        - name
+        - crn
+        - personType
+        - days
+        - bookingId
+        - roomId
+        - isSexualRisk
